# Optimize encoding of dictionaries