@fxamacker 2021-04-01
I found some code that requires better error handling.  I'll fix it and be more careful before opening my next PR.